requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
tavily-python>=0.3.3